Move command line functions to new file
